Move "show()" method to KGObject
