+++ fairgraph/base.py
@@ -21,6 +21,11 @@ try:
 except NameError:
     basestring = str
 import requests
+try:
+    from tabulate import tabulate
+    have_tabulate = True
+except ImportError:
+    have_tabulate = False
 from .errors import ResourceExistsError
 
 
@@ -471,6 +476,21 @@ class KGObject(with_metaclass(Registry, object)):
         else:
             for field in cls.fields:
                 field.strict_mode = value
+
+    def show(self, max_width=None):
+        if not have_tabulate:
+            raise Exception("You need to install the tabulate module to use the `show()` method")
+        data = [("id", self.id)] + [(field.name, getattr(self, field.name, None))
+                                    for field in self.fields]
+        if max_width:
+            value_column_width = max_width - max(len(item[0]) for item in data)
+            def fit_column(value):
+                strv = str(value)
+                if len(strv) > value_column_width:
+                    strv = strv[:value_column_width - 4] + " ..."
+                return strv
+            data = [(k, fit_column(v)) for k, v in data]
+        print(tabulate(data))
 
 
 class MockKGObject(KGObject):
